# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  price?: number | null;
  description: string;
  emoji: string;
  available: boolean;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @nullable */
  price?: number | null;
  description: string;
  emoji: string;
  available?: boolean;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  category?: string;
  /** @nullable */
  price?: number | null;
  description?: string;
  emoji?: string;
  available?: boolean;
}

export interface OrderItem {
  productId: number;
  name: string;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  price: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  delivered: 'delivered',
} as const;

export interface Order {
  id: number;
  studentName: string;
  phone: string;
  hostel: string;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
  createdAt: string;
}

export interface OrderInput {
  /** @minLength 1 */
  studentName: string;
  /** @minLength 5 */
  phone: string;
  /** @minLength 1 */
  hostel: string;
  /** @minItems 1 */
  items: OrderItem[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  delivered: 'delivered',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface DashboardSummary {
  totalOrders: number;
  pendingOrders: number;
  deliveredOrders: number;
  totalRevenue: number;
  totalProducts: number;
}