# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog operations
  - name: orders
    description: Student order operations
  - name: dashboard
    description: Admin dashboard operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      responses:
        "200":
          description: Product catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      security:
        - clerkSession: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      security:
        - clerkSession: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Product deleted
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      security:
        - clerkSession: []
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a student order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      security:
        - clerkSession: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary
      security:
        - clerkSession: []
      responses:
        "200":
          description: Order and catalog summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
components:
  securitySchemes:
    clerkSession:
      type: http
      scheme: bearer
      description: Clerk session authentication. Browser clients authenticate with the Clerk session cookie.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      required: [id, name, category, description, emoji, available, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        price: { type: ["number", "null"] }
        description: { type: string }
        emoji: { type: string }
        available: { type: boolean }
        createdAt: { type: string, format: date-time }
    ProductInput:
      type: object
      required: [name, category, description, emoji]
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        price: { type: ["number", "null"] }
        description: { type: string }
        emoji: { type: string }
        available: { type: boolean, default: true }
    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        price: { type: ["number", "null"] }
        description: { type: string }
        emoji: { type: string }
        available: { type: boolean }
    OrderItem:
      type: object
      required: [productId, name, quantity, price]
      properties:
        productId: { type: integer }
        name: { type: string }
        quantity: { type: integer, minimum: 1 }
        price: { type: number, minimum: 0 }
    Order:
      type: object
      required: [id, studentName, phone, hostel, items, total, status, createdAt]
      properties:
        id: { type: integer }
        studentName: { type: string }
        phone: { type: string }
        hostel: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        total: { type: number }
        status:
          type: string
          enum: [pending, delivered]
        createdAt: { type: string, format: date-time }
    OrderInput:
      type: object
      required: [studentName, phone, hostel, items]
      properties:
        studentName: { type: string, minLength: 1 }
        phone: { type: string, minLength: 5 }
        hostel: { type: string, minLength: 1 }
        items:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/OrderItem" }
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, delivered]
    DashboardSummary:
      type: object
      required: [totalOrders, pendingOrders, deliveredOrders, totalRevenue, totalProducts]
      properties:
        totalOrders: { type: integer }
        pendingOrders: { type: integer }
        deliveredOrders: { type: integer }
        totalRevenue: { type: number }
        totalProducts: { type: integer }