# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  studentName: string;
  phone: string;
  hostel: string;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
  createdAt: Date;
}