# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface OrderInput {
  /** @minLength 1 */
  studentName: string;
  /** @minLength 5 */
  phone: string;
  /** @minLength 1 */
  hostel: string;
  /** @minItems 1 */
  items: OrderItem[];
}