# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  price?: number | null;
  description: string;
  emoji: string;
  available: boolean;
  createdAt: Date;
}