# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  category?: string;
  /** @nullable */
  price?: number | null;
  description?: string;
  emoji?: string;
  available?: boolean;
}